Moved building complexity back into activator
