+++ pushHandler.go
@@ -3,11 +3,7 @@ package main
 import (
 	"net/http"
 	"os"
-
-	guuid "github.com/google/uuid"
-
 	"v9_deployment_manager/activator"
-	//	"v9_deployment_manager/database"
 	"v9_deployment_manager/log"
 	"v9_deployment_manager/worker"
 
@@ -65,60 +61,10 @@ func (h *pushHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	}
 	compID := worker.ComponentID{User: user, Repo: repo, Hash: hash}
 
-	// Get random tar name
-	tarName := guuid.New().String()
-	fullRepoName := compID.User + "/" + compID.Repo
-	// Get Repo Contents
-	log.Info.Println("Cloning " + compID.Repo + "...")
-	clonedPath, err := cloneRepo(fullRepoName)
-	if err != nil {
-		log.Error.Println("Error cloning repo:", err)
-		return
-	}
-	defer os.RemoveAll(clonedPath) // clean up
-
-	err = checkoutHead(clonedPath)
-	if err != nil {
-		log.Error.Println("Ding dong the witch is dead", err)
-	}
-	if compID.Hash == "HEAD" {
-		compID.Hash, err = getHash(clonedPath)
-		if err != nil {
-			log.Error.Println("Error getting hash from repo:", err)
-			return
-		}
-	}
-	// Build image
-	log.Info.Println("Building image from Dockerfile...")
-	err = buildImageFromDockerfile(tarName, clonedPath)
-	if err != nil {
-		log.Error.Println("Error building image from Dockerfile", err)
-		return
-	}
-
-	// Build and Zip Tar
-	log.Info.Println("Building and zipping tar...")
-	tarNameExt, err := buildAndZipTar(tarName)
-	if err != nil {
-		log.Error.Println("Failed to build and compress tar", err)
-		return
-	}
-	defer os.Remove("./" + tarNameExt)
-
-	// Send .tar to worker
-	log.Info.Println("SCP tar to worker...")
-	source := "./" + tarNameExt
-	destination := "/home/ubuntu/" + tarNameExt
-	err = scpToWorker(targetWorker.URL, source, destination, tarNameExt)
-	if err != nil {
-		log.Error.Println("Error copying to worker", err)
-		return
-	}
-
 	// Call deactivate to remove running component
 	worker.DeactivateComponentEverywhere(compID, h.workers)
 
-	err = h.activator.Activate(&compID, targetWorker, destination)
+	err = h.activator.Activate(&compID, targetWorker)
 	if err != nil {
 		log.Error.Println("Error activating worker", err)
 		return
